Move constants from `state` to `server`
